refactor bottom drawer
