+++ packages/uikit/src/components/Chart/PairPriceChart.tsx
@@ -195,7 +195,11 @@ export const SwapLineChart: React.FC<SwapLineChartNewProps> = ({
   return (
     <>
       {!chartCreated && <LineChartLoader />}
-      <div style={{ display: "flex", flex: 1, height: "100%" }} onMouseLeave={handleMouseLeave}>
+      <div
+        onPointerDownCapture={(event) => event.stopPropagation()}
+        style={{ display: "flex", flex: 1, height: "100%" }}
+        onMouseLeave={handleMouseLeave}
+      >
         <div style={{ flex: 1, maxWidth: "100%" }} ref={chartRef} id="swap-line-chart" {...rest} />
       </div>
     </>
