+++ packages/uikit/src/widgets/Modal/index.tsx
@@ -1,9 +1,10 @@
-export { default as Modal, MODAL_SWIPE_TO_CLOSE_VELOCITY, ModalWrapper } from "./Modal";
-export { default as ModalProvider } from "./ModalContext";
-export { default as useModal } from "./useModal";
-export * from "./styles";
-export type { ModalProps, InjectedProps as InjectedModalProps } from "./types";
-export * from "./ModalV2";
+export { default as BottomDrawer } from "./BottomDrawer/BottomDrawer";
+export { MODAL_SWIPE_TO_CLOSE_VELOCITY, default as Modal, ModalWrapper } from "./Modal";
 export { default as ModalActions } from "./ModalActions";
+export { default as ModalProvider } from "./ModalContext";
 export { default as ModalInput } from "./ModalInput";
+export * from "./ModalV2";
 export { default as NotEnoughTokensModal } from "./NotEnoughTokensModal";
+export * from "./styles";
+export type { InjectedProps as InjectedModalProps, ModalProps } from "./types";
+export { default as useModal } from "./useModal";
